API,Core: Move ScanReport to core module / extract TimerResult/CounterResult/ScanMetricsResult into own classes
The motivation behind moving `ScanReport` to `iceberg-core` is because we don't
actually need it in the `iceberg-api`, since `MetricsReporter` only
requires to have `MetricsReport` in the `iceberg-api` module.

The motivation for extracting all of those classes out of `ScanReport`
is so that we can reuse them in order to introduce an
`IncrementalScanReport` class, which slightly differs from `ScanReport`
around the snapshot (since it requires from/to snapshot IDs)
