+++ core/src/main/java/org/apache/iceberg/DeleteFileIndex.java
@@ -40,7 +40,7 @@ import org.apache.iceberg.expressions.ManifestEvaluator;
 import org.apache.iceberg.expressions.Projections;
 import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.io.FileIO;
-import org.apache.iceberg.metrics.ScanReport;
+import org.apache.iceberg.metrics.ScanMetrics;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.relocated.com.google.common.collect.Iterables;
@@ -364,7 +364,7 @@ class DeleteFileIndex {
     private PartitionSet partitionSet = null;
     private boolean caseSensitive = true;
     private ExecutorService executorService = null;
-    private ScanReport.ScanMetrics scanMetrics = ScanReport.ScanMetrics.noop();
+    private ScanMetrics scanMetrics = ScanMetrics.noop();
 
     Builder(FileIO io, Set<ManifestFile> deleteManifests) {
       this.io = io;
@@ -406,7 +406,7 @@ class DeleteFileIndex {
       return this;
     }
 
-    Builder scanMetrics(ScanReport.ScanMetrics newScanMetrics) {
+    Builder scanMetrics(ScanMetrics newScanMetrics) {
       this.scanMetrics = newScanMetrics;
       return this;
     }
