+++ core/src/main/java/org/apache/iceberg/ManifestReader.java
@@ -37,7 +37,7 @@ import org.apache.iceberg.io.CloseableGroup;
 import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.io.CloseableIterator;
 import org.apache.iceberg.io.InputFile;
-import org.apache.iceberg.metrics.ScanReport;
+import org.apache.iceberg.metrics.ScanMetrics;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableSet;
@@ -93,7 +93,7 @@ public class ManifestReader<F extends ContentFile<F>> extends CloseableGroup
   private Schema fileProjection = null;
   private Collection<String> columns = null;
   private boolean caseSensitive = true;
-  private ScanReport.ScanMetrics scanMetrics = ScanReport.ScanMetrics.noop();
+  private ScanMetrics scanMetrics = ScanMetrics.noop();
 
   // lazily initialized
   private Evaluator lazyEvaluator = null;
@@ -188,7 +188,7 @@ public class ManifestReader<F extends ContentFile<F>> extends CloseableGroup
     return this;
   }
 
-  ManifestReader<F> scanMetrics(ScanReport.ScanMetrics newScanMetrics) {
+  ManifestReader<F> scanMetrics(ScanMetrics newScanMetrics) {
     this.scanMetrics = newScanMetrics;
     return this;
   }
