+++ core/src/main/java/org/apache/iceberg/metrics/CounterResultParser.java
@@ -22,7 +22,6 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonNode;
 import java.io.IOException;
 import org.apache.iceberg.metrics.MetricsContext.Unit;
-import org.apache.iceberg.metrics.ScanReport.CounterResult;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.util.JsonUtil;
